hello-agent-assignment/simple-agent/pyproject.toml
[project]
name = "simple-agent"
version = "0.1.0"
description = "Minimal Hello Agent using OpenAI Agents SDK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai-agents",
    "python-dotenv",
]

